Changed encoder variable types to int
Gear reduction variables stayed as doubles because they were fractions by nature.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/RoverHardware.java
@@ -30,14 +30,14 @@ public class RoverHardware {
 
 
     // Encoder variables (REV Core Hex Motors)
-    public static final double COUNTS_PER_REV_CORE  = 288.0;
-    public static final double GEAR_REDUCTION_CORE  = 1.0; // No reduction on drive motors currently
-    public static final double COUNTS_PER_INCH_CORE = (COUNTS_PER_REV_CORE * GEAR_REDUCTION_CORE) / WHEEL_CIRCUMFERENCE; // To be confirmed
+    public static final int     COUNTS_PER_REV_CORE  = 288;
+    public static final double  GEAR_REDUCTION_CORE  = 1.0; // No reduction on drive motors currently
+    public static final int     COUNTS_PER_INCH_CORE = (int)((COUNTS_PER_REV_CORE * GEAR_REDUCTION_CORE) / WHEEL_CIRCUMFERENCE); // To be confirmed
 
     // Encoder variables (REV HD Hex Motors)
-    public static final double COUNTS_PER_REV_HD   = 1120.0;
-    public static final double GEAR_REDUCTION_HD   = 4.16; // 4.16 rotations of the motor shaft is 1 rotation of the drive pivot
-    public static final double COUNTS_PER_INCH_HD  = (COUNTS_PER_REV_HD * GEAR_REDUCTION_HD) / WHEEL_CIRCUMFERENCE; // To be confirmed
+    public static final int     COUNTS_PER_REV_HD   = 1120;
+    public static final double  GEAR_REDUCTION_HD   = 4.16; // 4.16 rotations of the motor shaft is 1 rotation of the drive pivot
+    public static final int     COUNTS_PER_INCH_HD  = (int)((COUNTS_PER_REV_HD * GEAR_REDUCTION_HD) / WHEEL_CIRCUMFERENCE); // To be confirmed
 
 
     // Pivot encoder variables
